requirements.txt
# YouTube Automation System — Python Dependencies
# Install with: pip install -r requirements.txt

# ─── Anthropic Claude ──────────────────────
anthropic>=0.39.0

# ─── YouTube / Google APIs ─────────────────
google-api-python-client>=2.150.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0

# ─── Web Search & Scraping ─────────────────
requests>=2.32.0
beautifulsoup4>=4.12.0
aiohttp>=3.10.0
pytrends>=4.9.2
serper-dev>=0.1.0

# ─── TTS / Audio ───────────────────────────
edge-tts>=7.0.0
soundfile>=0.12.1
numpy>=1.26.0
pydub>=0.25.1

# ─── Image Generation ──────────────────────
Pillow>=10.4.0

# ─── Scheduling & Async ────────────────────
APScheduler>=3.10.4
asyncio-throttle>=1.0.2

# ─── Config & Utilities ────────────────────
python-dotenv>=1.0.1
rich>=13.8.0

# ─── Optional: Kokoro TTS (best quality) ───
# Install separately: pip install kokoro-onnx
# kokoro-onnx>=0.4.0
